Revert "Revert "ResourceTypeParser""
This reverts commit 41ab65875d8b815041bf06b8b964c8bf4fc5cb77.
